Hi!

First of all, thank you very much for writing this well structured library!

With inspiration from https://github.com/catid/leopard/ I wrote a new `engine` that leverages AVX2 instructions.
On an AMD Ryzen 5 3600, encoding is speed up by up to 1700%, and decoding up to about 1100%. Ex:
```
main/ReedSolomonEncoder/1000:1000
                        time:   [462.20 µs 462.81 µs 463.49 µs]
                        thrpt:  [4.1152 GiB/s 4.1212 GiB/s 4.1267 GiB/s]
                 change:
                        time:   [-94.288% -94.173% -94.088%] (p = 0.00 < 0.05)
                        thrpt:  [+1591.5% +1616.2% +1650.6%]
                        Performance has improved.

main/ReedSolomonDecoder/1000:1000 (1%)
                        time:   [1.5372 ms 1.5431 ms 1.5492 ms]
                        thrpt:  [1.2312 GiB/s 1.2361 GiB/s 1.2408 GiB/s]
                 change:
                        time:   [-91.384% -91.308% -91.237%] (p = 0.00 < 0.05)
                        thrpt:  [+1041.1% +1050.4% +1060.6%]
                        Performance has improved.

main/ReedSolomonDecoder/1000:1000 (100%)
                        time:   [1.5545 ms 1.5630 ms 1.5724 ms]
                        thrpt:  [1.2130 GiB/s 1.2203 GiB/s 1.2270 GiB/s]
                 change:
                        time:   [-91.580% -91.525% -91.467%] (p = 0.00 < 0.05)
                        thrpt:  [+1071.9% +1079.9% +1087.6%]
                        Performance has improved.
```
Full run: https://pastebin.com/m8AcP8T5