@AndersTrier 2023-11-04
Some relevant documentation: https://doc.rust-lang.org/core/arch/index.html

Here's my suggestion to how choosing this engine should behave:
By default, at runtime we should detect if the CPU supports AVX2 (this is what `is_x86_feature_detected!("avx2")` does), and fallback to using the `NoSimd` engine.

And since the `Avx2` engine uses `unsafe`, I think there should be a way to turn it off.
This is currently implemented as a `feature` called `avx2` which is enabled by default. I.e Cargo.toml contains:
```
[features]
default = ["avx2"]
avx2 = []
```
An alternative would be to have `no_unsafe` as an opt-in feature. Which one do you prefer?

Anyways, essentially what I would like is something like this in `engine.rs`:
```rust
#[cfg(all(feature = "avx2", any(target_arch = "x86", target_arch = "x86_64")))]
pub type DefaultEngine = if is_x86_feature_detected!("avx2") { Avx2 } else { NoSimd };
```
But that doesn't work for a type alias.

I tried with something like
```rust
pub enum DefaultEngine {}

impl DefaultEngine {
    pub fn new() -> impl Engine {
#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
        {
            if is_x86_feature_detected!("avx2") {
                return Avx2::new();
            }
        }
        NoSimd::new()
    }
}
```
But that seems to require a modification of `HighRate` and `LowRate`, and before I attempted that, I figured it would be good with some feedback.