@AndersTrier 2023-11-04
We could totally avoid `unsafe` by using https://docs.rs/safe_arch/latest/safe_arch/
but as far as I understand, that would require every user of the `reed-solomon-16` library to compile their code with `-C target-feature=+avx2`.

Note: here's a footgun. By default, `rustc` only targets `SSE` but will happily compile the AVX2 intrinsics without you specifying `-C target-feature=+avx2`. The AVX2 intrinsics will then be "emulated" using SSE (also on a CPU with support for AVX2!). This makes `Avx2` much slower than `NoSimd`.

That is why most of the methods are annotated with `#[target_feature(enable = "avx2")]` which tells the compiler to assume that AVX2 instructions are available, but that in turn requires the method to be marked as `unsafe`, even though it doesn't actually need unsafe privileges.

We could drop the privileges again by using another layer of abstraction. I.e:
```
impl mul() -> mul_unsafe_avx2enabled() -> mul_safe()
```
It adds quite a few "dummy" functions, but would move a lot of the code out of unsafe. Would you like to see an example of what that would look like?