@malaire 2023-11-04
Nice work! Unfortunately I've read that SIMD code of Leopard may have patent restrictions.

For example this comment at [FastECC](https://github.com/Bulat-Ziganshin/FastECC):
> Moreover, FastECC is free from patent restrictions that has any fast RS codec employing PSHUFB (i.e. SSSE3) including Leopard.

I havn't been able to verify this (nor do I have enough knowledge to verify patent claims), but I've seen some evidence suggesting that this may be true. I can't take any risks concerning patents and so I don't think I can include any SIMD code for now.

ps. I don't have any experience on Rust SIMD so currently I also can't review code like this. I've been meaning to learn that, but because of the patent problem I havn't even started on this.