@AndersTrier 2023-11-06
> Nice work!

Thanks!

> I've read that SIMD code of Leopard may have patent restrictions. I havn't been able to verify this (nor do I have enough knowledge to verify patent claims), but I've seen some evidence suggesting that this may be true. I can't take any risks concerning patents and so I don't think I can include any SIMD code for now.

Think I found the patent you are referring to:
http://www.google.com/patents/US8683296

A break down of the patent can be found here: https://erasurecodepatents.wordpress.com/

It looks like StreamScale is having at least some success enforcing it: https://www.datanami.com/2023/10/16/cloudera-hit-with-240-million-judgement-over-erasure-coding/

But it sounds like Cloudera is going to appeal though:
> We disagree with and are extremely disappointed by the decision. We remain steadfast in our belief in our defenses in the case, and we intend to challenge the judgment.

(Btw, the US District Court for the Western District of Texas is know for beeing very patent friendly: https://www.eff.org/deeplinks/2014/07/why-do-patent-trolls-go-texas-its-not-bbq )

Anyways, as that patent is only a risk for companies doing business in the US, how about making this code opt-in, and including a warning in the readme?

There are many other open source projects that include variations of this code:
@klauspost's Go library: https://github.com/klauspost/reedsolomon/
Red Hat Ceph: https://github.com/ceph/ceph/tree/main/src/erasure-code/jerasure
The Linux Kernel: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/lib/raid6/avx2.c
Intel's Intelligent Storage Acceleration Library: https://github.com/intel/isa-l