@malaire 2023-11-11
So they have successfully enforced it against a company and got an open-source project (GF-Complete & Jerasure) to quit.

I can't take any risks with patent like that. Also my interest in this algorithm is longterm and I have no problem waiting until 2032 when that patent expires before adding SIMD.

p.s. It's always possible to implement AVX 2 engine in a separate crate, if you or someone else thinks this patent isn't a problem.