@AndersTrier 2023-11-12
That's totally fair. I'll fork and submit a separate crate