@AndersTrier 2023-11-25
For anyone interested, I've submitted a fork under the name `reed-solomon-simd`: https://crates.io/crates/reed-solomon-simd

On AArch64 it uses the Neon SIMD instructions, and on x86(-64) either AVX2 or SSSE3 is used. The best implementation is chosen at runtime with fallback to plain Rust.